Refactor tests : decouple rdd integration tests with basic tensor unit tests
